# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
# host: localhost:5040
host: 52.39.212.226:5054
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/getUsersPermission:
    # binds a127 app logic to a route
    x-swagger-router-controller: permissions_ctrl
    get:
      tags: ['Permission Module']
      description: Get Users Permission
      summary: Get Users Permission
      # used as the method name of the controller
      operationId: getUserPermissions  
      parameters:    
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
...
  /api/permissions:
    # binds a127 app logic to a route
    x-swagger-router-controller: permissions_ctrl
    get:
      tags: ['permissions']
      description: Get permissions
      summary: Get permissions
      # used as the method name of the controller
      operationId: getPermissions
      parameters:
        - name: page
          in: query
          description: page
          required: false
          type: string
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string       
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
    # post:
    #   tags: ['permissions']
    #   description: Create permissions
    #   summary: Create permissions
    #   # used as the method name of the controller
    #   operationId: createPermission
    #   parameters:
    #     - name: body
    #       in: body
    #       description: permission infomation
    #       schema:
    #        $ref: '#/definitions/createPermission'
    #       required: true
    #   responses:
    #     "200":
    #       description: Success
    #       schema:
    #         # a pointer to a definition
    #         $ref: "#/definitions/appResponse"
    #     # responses may fall through to errors
    #     default:
    #       description: Error
    #       schema:
    #         $ref: "#/definitions/ErrorResponse"  
...
  /api/addkupatCholim:
    # binds a127 app logic to a route
    x-swagger-router-controller: kupat_ctrl
    post:
      tags: ['Kupat Module']
      description: add kupat_cholim
      summary:  add kupat_cholim
      # used as the method name of the controller
      operationId: addkupatCholim
      parameters:
        - name: body
          in: body
          description:  add kupat_cholim
          schema:
           $ref: '#/definitions/addkupatCholim'
          required: true       
        # - name: Authorization
        #   in: header
        #   description: Authorization
        #   required: true
        #   type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getkupatCholim:
    # binds a127 app logic to a route
    x-swagger-router-controller: kupat_ctrl
    get:
      tags: ['Kupat Module']
      description: get kupat_cholim
      summary:  get kupat_cholim
      # used as the method name of the controller
      operationId: getkupatCholim      
      # parameters:
      #   - name: Getuser
      #     in: header
      #     description: Getuser
      #     required: true
      #     type: string                     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getkupatCholimbyId/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: kupat_ctrl
    get:
      tags: ['Kupat Module']
      description: get kupat_cholim by id
      summary: get kupat_cholim by id
      # used as the method name of the controller
      operationId: getkupatCholimbyId
      parameters:
        - name: id
          in: path
          description: get kupat_cholim by id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"           
  /api/updatekupatCholim/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: kupat_ctrl
    put:
      tags: ['Kupat Module']
      description: update kupat_cholim
      summary:  Update kupat_cholim
      operationId: updatekupatCholim
      parameters:
        - name: id
          in: path
          description: kupat_cholim id
          required: true
          type: string       
        - name: body
          in: body
          description: Body
          required: false
          schema:
            # a pointer to a definition
            $ref: "#/definitions/updatekupatCholim"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/deletekupatCholimbyId/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: kupat_ctrl
    delete:
      tags: ['Kupat Module']
      description: delete kupat_cholim
      summary:  delete kupat_cholim
      # used as the method name of the controller
      operationId: deletekupatCholimbyId
      parameters:
        - name: id
          in: path
          description: delete kupat_cholim by id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
...
definitions:                         
...
  addkupatCholim:
    type: object
    properties:
      name: 
        type: string
      price:
        type: number
      email: 
        type: string
      weekendPriceSet:
        type: array
        items:
          type: object
          properties:
           title:
            type: string
           price: 
            type: number
           dateRange:
            type: string
  updatekupatCholim:
    type: object
    properties:
      name: 
        type: string
      price:
        type: number
      email: 
        type: string
      weekendPrice:
        type: number
      weekendDayRange:
        type: string
      weekendPriceSet:
        type: array
        items:
          type: object
          properties:
           title:
            type: string
           price: 
            type: number
           dateRange:
            type: string
...
  appResponse:
    # required:
    #   - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  # by sunny on 15-april-2017
...